+++ pkg/api/dtos/plugin_bundle.go
@@ -1,0 +1,8 @@
+package dtos
+
+type PluginBundle struct {
+	Type     string                 `json:"type"`
+	Enabled  bool                   `json:"enabled"`
+	Module   string                 `json:"module"`
+	JsonData map[string]interface{} `json:"jsonData"`
+}