+++ public/app/app.js
@@ -2,19 +2,18 @@ define([
   'angular',
   'jquery',
   'lodash',
+  'app/core/config',
   'require',
-  'config',
   'bootstrap',
   'angular-route',
   'angular-sanitize',
   'angular-strap',
   'angular-dragdrop',
   'angular-ui',
-  'extend-jquery',
   'bindonce',
   'app/core/core',
 ],
-function (angular, $, _, appLevelRequire) {
+function (angular, $, _, config, appLevelRequire) {
   "use strict";
 
   var app = angular.module('grafana', []);
@@ -37,6 +36,8 @@ function (angular, $, _, appLevelRequire) {
     } else {
       _.extend(module, register_fns);
     }
+    // push it into the apps dependencies
+    apps_deps.push(module.name);
     return module;
   };
 
@@ -66,16 +67,15 @@ function (angular, $, _, appLevelRequire) {
     var module_name = 'grafana.'+type;
     // create the module
     app.useModule(angular.module(module_name, []));
-    // push it into the apps dependencies
-    apps_deps.push(module_name);
   });
 
-  var preBootRequires = [
-    'app/services/all',
-    'app/features/all',
-    'app/controllers/all',
-    'app/components/partials',
-  ];
+  var preBootRequires = ['app/features/all'];
+  var pluginModules = config.bootData.pluginModules || [];
+
+  // add plugin modules
+  for (var i = 0; i < pluginModules.length; i++) {
+    preBootRequires.push(pluginModules[i]);
+  }
 
   app.boot = function() {
     require(preBootRequires, function () {