+++ public/app/features/templating/editorCtrl.js
@@ -12,7 +12,7 @@ function (angular, _) {
     var replacementDefaults = {
       type: 'query',
       datasource: null,
-      refresh_on_load: false,
+      refresh: false,
       name: '',
       options: [],
       includeAll: false,
@@ -94,6 +94,13 @@ function (angular, _) {
         $scope.current.type = 'query';
         $scope.current.allFormat = 'glob';
       }
+    };
+
+    $scope.duplicate = function(variable) {
+      $scope.current = angular.copy(variable);
+      $scope.variables.push($scope.current);
+      $scope.current.name = 'copy_of_'+variable.name;
+      $scope.updateSubmenuVisibility();
     };
 
     $scope.update = function() {