+++ public/app/plugins/datasource/elasticsearch/datasource.js
@@ -2,7 +2,7 @@ define([
   'angular',
   'lodash',
   'moment',
-  'kbn',
+  'app/core/utils/kbn',
   './query_builder',
   './index_pattern',
   './elastic_response',
@@ -19,13 +19,17 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
     function ElasticDatasource(datasource) {
       this.type = 'elasticsearch';
       this.basicAuth = datasource.basicAuth;
+      this.withCredentials = datasource.withCredentials;
       this.url = datasource.url;
       this.name = datasource.name;
       this.index = datasource.index;
       this.timeField = datasource.jsonData.timeField;
+      this.esVersion = datasource.jsonData.esVersion;
       this.indexPattern = new IndexPattern(datasource.index, datasource.jsonData.interval);
+      this.interval = datasource.jsonData.timeInterval;
       this.queryBuilder = new ElasticQueryBuilder({
-        timeField: this.timeField
+        timeField: this.timeField,
+        esVersion: this.esVersion,
       });
     }
 
@@ -36,8 +40,10 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
         data: data
       };
 
-      if (this.basicAuth) {
+      if (this.basicAuth || this.withCredentials) {
         options.withCredentials = true;
+      }
+      if (this.basicAuth) {
         options.headers = {
           "Authorization": this.basicAuth
         };
@@ -60,17 +66,18 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
         });
     };
 
-    ElasticDatasource.prototype.annotationQuery = function(annotation, rangeUnparsed) {
-      var range = {};
+    ElasticDatasource.prototype.annotationQuery = function(options) {
+      var annotation = options.annotation;
       var timeField = annotation.timeField || '@timestamp';
       var queryString = annotation.query || '*';
       var tagsField = annotation.tagsField || 'tags';
       var titleField = annotation.titleField || 'desc';
       var textField = annotation.textField || null;
 
+      var range = {};
       range[timeField]= {
-        from: rangeUnparsed.from,
-        to: rangeUnparsed.to,
+        from: options.range.from.valueOf(),
+        to: options.range.to.valueOf(),
       };
 
       var queryInterpolated = templateSrv.replace(queryString);
@@ -82,9 +89,20 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
         "size": 10000
       };
 
-      return this._request('POST', annotation.index + '/_search', data).then(function(results) {
+      var header = {search_type: "query_then_fetch", "ignore_unavailable": true};
+
+      // old elastic annotations had index specified on them
+      if (annotation.index) {
+        header.index = annotation.index;
+      } else {
+        header.index = this.indexPattern.getIndexList(options.range.from, options.range.to);
+      }
+
+      var payload = angular.toJson(header) + '\n' + angular.toJson(data) + '\n';
+
+      return this._post('_msearch', payload).then(function(res) {
         var list = [];
-        var hits = results.data.hits.hits;
+        var hits = res.responses[0].hits.hits;
 
         var getFieldFromSource = function(source, fieldName) {
           if (!fieldName) { return; }
@@ -95,7 +113,7 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
           for (var i = 0; i < fieldNames.length; i++) {
             fieldValue = fieldValue[fieldNames[i]];
             if (!fieldValue) {
-              console.log('could not find field in annotatation: ', fieldName);
+              console.log('could not find field in annotation: ', fieldName);
               return '';
             }
           }
@@ -141,8 +159,8 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
       });
     };
 
-    ElasticDatasource.prototype.getQueryHeader = function(timeFrom, timeTo) {
-      var header = {search_type: "count", "ignore_unavailable": true};
+    ElasticDatasource.prototype.getQueryHeader = function(searchType, timeFrom, timeTo) {
+      var header = {search_type: searchType, "ignore_unavailable": true};
       header.index = this.indexPattern.getIndexList(timeFrom, timeTo);
       return angular.toJson(header);
     };
@@ -152,20 +170,27 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
       var target;
       var sentTargets = [];
 
-      var header = this.getQueryHeader(options.range.from, options.range.to);
-
       for (var i = 0; i < options.targets.length; i++) {
         target = options.targets[i];
-        if (target.hide) {return;}
+        if (target.hide) {continue;}
 
-        var esQuery = angular.toJson(this.queryBuilder.build(target));
+        var queryObj = this.queryBuilder.build(target);
+        var esQuery = angular.toJson(queryObj);
         var luceneQuery = angular.toJson(target.query || '*');
         // remove inner quotes
         luceneQuery = luceneQuery.substr(1, luceneQuery.length - 2);
         esQuery = esQuery.replace("$lucene_query", luceneQuery);
 
-        payload += header + '\n' + esQuery + '\n';
+        var searchType = queryObj.size === 0 ? 'count' : 'query_then_fetch';
+        var header = this.getQueryHeader(searchType, options.range.from, options.range.to);
+        payload +=  header + '\n';
+
+        payload += esQuery + '\n';
         sentTargets.push(target);
+      }
+
+      if (sentTargets.length === 0) {
+        return $q.when([]);
       }
 
       payload = payload.replace(/\$interval/g, options.interval);
@@ -173,7 +198,7 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
       payload = payload.replace(/\$timeTo/g, options.range.to.valueOf());
       payload = templateSrv.replace(payload, options.scopedVars);
 
-      return this._post('/_msearch?search_type=count', payload).then(function(res) {
+      return this._post('_msearch', payload).then(function(res) {
         return new ElasticResponse(sentTargets, res).getTimeSeries();
       });
     };
@@ -217,7 +242,7 @@ function (angular, _, moment, kbn, ElasticQueryBuilder, IndexPattern, ElasticRes
 
     ElasticDatasource.prototype.getTerms = function(queryDef) {
       var range = timeSrv.timeRange();
-      var header = this.getQueryHeader(range.from, range.to);
+      var header = this.getQueryHeader('count', range.from, range.to);
       var esQuery = angular.toJson(this.queryBuilder.getTermsQuery(queryDef));
 
       esQuery = esQuery.replace("$lucene_query", queryDef.query || '*');