+++ public/app/plugins/datasource/prometheus/datasource.js
@@ -21,7 +21,9 @@ function (angular, _, moment, dateMath) {
       this.name = datasource.name;
       this.supportMetrics = true;
       this.url = datasource.url;
+      this.directUrl = datasource.directUrl;
       this.basicAuth = datasource.basicAuth;
+      this.withCredentials = datasource.withCredentials;
       this.lastErrors = {};
     }
 
@@ -31,8 +33,10 @@ function (angular, _, moment, dateMath) {
         method: method
       };
 
-      if (this.basicAuth) {
+      if (this.basicAuth || this.withCredentials) {
         options.withCredentials = true;
+      }
+      if (this.basicAuth) {
         options.headers = {
           "Authorization": this.basicAuth
         };
@@ -47,6 +51,7 @@ function (angular, _, moment, dateMath) {
       var end = getPrometheusTime(options.range.to, true);
 
       var queries = [];
+      options = _.clone(options);
       _.each(options.targets, _.bind(function(target) {
         if (!target.expr || target.hide) {
           return;
@@ -57,7 +62,13 @@ function (angular, _, moment, dateMath) {
 
         var interval = target.interval || options.interval;
         var intervalFactor = target.intervalFactor || 1;
-        query.step = this.calculateInterval(interval, intervalFactor);
+        target.step = query.step = this.calculateInterval(interval, intervalFactor);
+        var range = Math.ceil(end - start);
+        // Prometheus drop query if range/step > 11000
+        // calibrate step if it is too big
+        if (query.step !== 0 && range / query.step > 11000) {
+          target.step = query.step = Math.ceil(range / 11000);
+        }
 
         queries.push(query);
       }, this));
@@ -95,17 +106,7 @@ function (angular, _, moment, dateMath) {
     };
 
     PrometheusDatasource.prototype.performTimeSeriesQuery = function(query, start, end) {
-      var url = '/api/v1/query_range?query=' + encodeURIComponent(query.expr) + '&start=' + start + '&end=' + end;
-
-      var step = query.step;
-      var range = Math.floor(end - start);
-      // Prometheus drop query if range/step > 11000
-      // calibrate step if it is too big
-      if (step !== 0 && range / step > 11000) {
-        step = Math.floor(range / 11000);
-      }
-      url += '&step=' + step;
-
+      var url = '/api/v1/query_range?query=' + encodeURIComponent(query.expr) + '&start=' + start + '&end=' + end + '&step=' + query.step;
       return this._request('GET', url);
     };
 
@@ -113,11 +114,9 @@ function (angular, _, moment, dateMath) {
       var url = '/api/v1/label/__name__/values';
 
       return this._request('GET', url).then(function(result) {
-        var suggestData = _.filter(result.data.data, function(metricName) {
-          return metricName.indexOf(query) !==  1;
+        return _.filter(result.data.data, function (metricName) {
+          return metricName.indexOf(query) !== 1;
         });
-
-        return suggestData;
       });
     };
 
@@ -211,7 +210,7 @@ function (angular, _, moment, dateMath) {
         sec = 1;
       }
 
-      return Math.floor(sec * intervalFactor) + 's';
+      return Math.ceil(sec * intervalFactor);
     };
 
     function transformMetricData(md, options) {
@@ -220,8 +219,20 @@ function (angular, _, moment, dateMath) {
 
       metricLabel = createMetricLabel(md.metric, options);
 
-      dps = _.map(md.values, function(value) {
-        return [parseFloat(value[1]), value[0] * 1000];
+      var stepMs = parseInt(options.step) * 1000;
+      var lastTimestamp = null;
+      _.each(md.values, function(value) {
+        var dp_value = parseFloat(value[1]);
+        if (_.isNaN(dp_value)) {
+          dp_value = null;
+        }
+
+        var timestamp = value[0] * 1000;
+        if (lastTimestamp && (timestamp - lastTimestamp) > stepMs) {
+          dps.push([null, lastTimestamp + stepMs]);
+        }
+        lastTimestamp = timestamp;
+        dps.push([dp_value, timestamp]);
       });
 
       return { target: metricLabel, datapoints: dps };